add dispatch to actions
